Check for start_time column before removing it in bookings

+++ db/migrate/20240426095412_change_time_frame_type_in_booking.rb
@@ -1,6 +1,7 @@
 class ChangeTimeFrameTypeInBooking < ActiveRecord::Migration[7.1]
   def change
-    remove_column :bookings, :start_time, :datetime
-    add_column :bookings, :start_time, :datetime
+    if column_exists?(:bookings, :start_time)
+      remove_column :bookings, :start_time, :datetime
+    end
   end
 end
